pyproject.toml
[project]
name = "aiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain==1.2.7",
    "langchain-anthropic==1.3.1",
    "langchain-google-genai==4.2.0",
    "langchain-groq==1.1.2",
    "langchain-ollama==1.0.1",
    "langchain-openai==1.1.7",
    "python-dotenv==1.1.0",
    "rich==13.9.4",
]
